Orbit intersects.

+++ Limnova/src/Math/MathConstants.h
@@ -5,10 +5,10 @@ namespace Limnova
 {
 
     constexpr double PI         = 3.1415926535897932384626433832795028841971693993751058209;
-    constexpr float PI2         = PI * 2.0;
-    constexpr float PIover2     = PI * 0.5;
-    constexpr float PIover4     = PI * 0.25;
-    constexpr float PIover8     = PI * 0.125;
+    constexpr double PI2         = PI * 2.0;
+    constexpr double PIover2     = PI * 0.5;
+    constexpr double PIover4     = PI * 0.25;
+    constexpr double PIover8     = PI * 0.125;
 
     constexpr float PIf         = (float)PI;
     constexpr float PI2f        = (float)PI2;
